4ta entrega

+++ EasyTrufi.Core/Interfaces/IBaseRepository.cs
@@ -10,9 +10,9 @@ namespace EasyTrufi.Core.Interfaces
     public interface IBaseRepository<T> where T : BaseEntity
     {
         Task<IEnumerable<T>> GetAll();
-        Task<T> GetById(int id);
+        Task<T> GetById(long id);
         Task Add(T entity);
         Task Update(T entity);
-        Task Delete(int id);
+        Task Delete(long id);
     }
 }
